+++ tests/Nakama.Tests/LinkUnlinkTest.cs
@@ -69,7 +69,7 @@ namespace Nakama.Tests.Api
             var account = await _client.GetAccountAsync(session);
 
             Assert.NotNull(account);
-            Assert.Equal(account.Devices.Count(d => d.Id == deviceid), 1);
+            Assert.Equal(1, account.Devices.Count(d => d.Id == deviceid));
             Assert.Equal(customid, account.CustomId);
         }
 
@@ -84,7 +84,7 @@ namespace Nakama.Tests.Api
             var account = await _client.GetAccountAsync(session);
 
             Assert.NotNull(account);
-            Assert.Equal(account.Devices.Count(d => d.Id == deviceid), 1);
+            Assert.Single(account.Devices, d => d.Id == deviceid);
             Assert.Null(account.CustomId);
         }
 
@@ -172,7 +172,7 @@ namespace Nakama.Tests.Api
             var account = await _client.GetAccountAsync(session);
 
             Assert.NotNull(account);
-            Assert.Equal(account.Devices.Count(d => d.Id == deviceid), 0);
+            Assert.Equal(0, account.Devices.Count(d => d.Id == deviceid));
         }
 
         [Fact(Timeout = TestsUtil.TIMEOUT_MILLISECONDS)]